edit event page and events map page to resolve migration erorrs
