+++ AppDev/AppDev/Models/Event.swift
@@ -6,3 +6,15 @@
 //
 
 import Foundation
+import MapKit
+
+struct Event: Identifiable {
+    let id = UUID()
+    let title: String
+    let date: String
+    let location: String
+    let coordinate: CLLocationCoordinate2D
+    let imageUrl: String
+    let attendees: Int
+    let distance: String
+}
